nncs/requirements.txt
torch==1.11.0
numpy==1.22.3
torchvision==0.12.0
easydict
networkx
onnx2torch
onnxruntime==1.13.1
pydot
